requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.9
python-dotenv==1.0.1

chromadb==1.0.20
sentence-transformers==5.1.0
rank-bm25==0.2.2

pdfplumber==0.11.7
python-docx==1.2.0
python-pptx==1.0.2
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.3.0
